chore: separate constants in /lib into separate constants files / add default storefront api version
