+++ app/lib/constants/collection.ts
@@ -1,0 +1,3 @@
+export const COLOR_OPTION_NAME = 'Color' as const;
+
+export const PRICE_FILTER_ID = 'filter.v.price' as const;
