+++ app/lib/constants/navigation.ts
@@ -1,0 +1,3 @@
+export const PROMOBAR_HEIGHT_MOBILE = 48; /* Ensure this number (px) is equivalent to --promobar-height-mobile in app.css */
+
+export const PROMOBAR_HEIGHT_DESKTOP = 48; /* Ensure this number (px) equivalent to --promobar-height-desktop in app.css */
