+++ app/lib/constants/product.ts
@@ -1,0 +1,5 @@
+import type {AspectRatio} from '../types';
+
+/* Ensure updating this ratio as needed. Required format is 'width/height' */
+export const PRODUCT_IMAGE_ASPECT_RATIO: AspectRatio =
+  '3/4'; /* Ensure this is equivalent to product-image-aspect-ratio in app.css */
